+++ packages/shared/src/components/recruiter/layout/RecruiterLayout.tsx
@@ -1,11 +1,11 @@
 import type { ReactElement, ReactNode } from 'react';
-import React, { useContext } from 'react';
+import React from 'react';
 import classNames from 'classnames';
 import { InAppNotificationElement } from '../../notifications/InAppNotification';
 import { PromptElement } from '../../modals/Prompt';
 import { Sidebar } from './Sidebar';
 import Toast from '../../notifications/Toast';
-import SettingsContext from '../../../contexts/SettingsContext';
+import { useSettingsContext } from '../../../contexts/SettingsContext';
 
 export type RecruiterSelfServeLayoutProps = {
   children: ReactNode | undefined;
@@ -18,7 +18,7 @@ export const RecruiterLayout = ({
   children,
   className,
 }: RecruiterSelfServeLayoutProps): ReactElement => {
-  const { autoDismissNotifications } = useContext(SettingsContext);
+  const { autoDismissNotifications } = useSettingsContext();
 
   return (
     <div className="antialiased">
